Rename file: metadata/index.js -> aside/metadata.js
