+++ src/components/article-page.js
@@ -2,7 +2,7 @@ import { ThemeProvider } from 'styled-components'
 import Body from './body'
 import get from 'lodash/get'
 import LeadingBlock from './leading-block'
-import Metadata from './metadata'
+import Metadata from './aside/metadata'
 import predefinedPropTypes from '../constants/prop-types'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
